http://jira.jboss.org/jira/browse/JBMESSAGING-782
We need to increase the number of times a socket client invoker tries to write on an existing
socket and then retires creating the socket (see MicroSocketClientInvoker.transport()), 
because otherwise the invoker will fail writing on timed out sockets, which is a normal
condition, and then it won't try recreating the socket, which it should.
We can get away with numberOfCallRetries = 2.


git-svn-id: http://anonsvn.jboss.org/repos/hornetq/trunk@2068 7e4d6552-b66c-0410-90c2-bdaa8f2d6dd4