http://jira.jboss.org/jira/browse/JBMESSAGING-898
Summary of changes:

Testsuite:
Added a testsuite on graveyard (a test that would kill server0... what would invalidate the test run)
Added ClusteredConnectionFactoryTest, and also added poisoning support for ConnectionFactory
   (Adding poisoning required to expose serverPeer on ServerConnectionFactoryEndpoint)

Implementation:
Protected the connection creation in a loop, and doing retries until a max is reached.
Testing connection/IO failures on getClientAOPConfig, retrying on any delegates. (notice that a failure on getClientAOPConfig is expected to be a rare event, as this is executed only once)

git-svn-id: http://anonsvn.jboss.org/repos/hornetq/trunk@2466 7e4d6552-b66c-0410-90c2-bdaa8f2d6dd4