Fixed runaway clustering server. This is what happened:
The ant script used to pre-stage 4 servers (0, 1, 2, 3) for clustering tests.
While running ClusteredConnectionFactoryTest, server 0 (the one that runs the
RMI registry) would be killed, so even if restarted later to give a 
RMI registry to the test, server 3, which is not used during the test, will
become stranded, and the cleanup sequence won't stop it.

Solution: don't prestage servers. Let the test start the servers it needs.


git-svn-id: http://anonsvn.jboss.org/repos/hornetq/trunk@2488 7e4d6552-b66c-0410-90c2-bdaa8f2d6dd4